refactor(config/grafana): split config.go by API group
Split the monolithic config.go (~676 lines) into per-API-group files
matching the Crossplane API surface defined in config/groups.go:

- alerting.go: notification_policy, contact_point, rule_group
- cloud.go: cloud_access_policy, cloud_access_policy_token,
  cloud_plugin_installation, cloud_stack,
  cloud_stack_service_account, cloud_stack_service_account_token
- enterprise.go: data_source_permission, report, role,
  role_assignment, role_assignment_item, team_external_group
- oss.go: annotation, dashboard, dashboard_public,
  dashboard_permission, folder, folder_permission, library_panel,
  service_account_permission, service_account_token, team,
  apps_alertenrichment_alertenrichment_v1beta1
- sm.go: synthetic_monitoring_installation,
  synthetic_monitoring_check_alerts
- ml.go: machine_learning_job, machine_learning_outlier_detector
- slo.go: slo
- k6.go: k6_project_limits, k6_load_test
- oncall.go: oncall_escalation, oncall_integration, oncall_route,
  oncall_schedule

Merges the duplicate notification_policy configurator and removes the
exported ConfigureOnCallRefsAndSelectors function, inlining its content
into the per-group files. The apps_alertenrichment resource moves from
oncall to oss where it belongs per its category.
